登录逻辑

+++ server/src/main/java/icu/qihangduan/dachuang_server/pojo/User.java
@@ -1,7 +1,9 @@
 package icu.qihangduan.dachuang_server.pojo;
 
 import com.baomidou.mybatisplus.annotation.FieldFill;
+import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
+import com.baomidou.mybatisplus.annotation.TableId;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -19,7 +21,7 @@ import java.util.Date;
 @AllArgsConstructor
 @NoArgsConstructor
 public class User {
-    private Integer id;
+    private Long id;
     private String email;
     private String username;
     private String password;
